re-order and re-format code; such as, creating helper files, and modularizing code accordingly
