+++ src/backend-src/routes/api/index.ts
@@ -5,6 +5,6 @@ const router = require('express').Router();
 const usersRouter = require('./users.js');
 const projectsRouter = require('./projects');
 
-router.use('/projects', projectsRouter)
+router.use('/projects', projectsRouter);
 router.use('/users', usersRouter);
 module.exports = router;
