+++ src/backend-src/routes/api/users.ts
@@ -1,63 +1,27 @@
 // packages
 import express, {
-    Application, 
     Request,
     Response, 
     NextFunction
 } from 'express';
-import { ExpError } from '../../custom-types';
-const asyncHandler = require('express-async-handler');
-const { check } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validation');
+import { ExpError } from '../../types/d';
 
-const { setTokenCookie, restoreUser, requireAuth } = require('../../utils/auth');
+const {validateLogin, validateSignup} = require('../../utils/validation');
+const asyncHandler = require('express-async-handler');
+
+const { setTokenCookie, restoreUser } = require('../../utils/auth');
 const { User } = require('../../db/models');
 
 
-const bookmarksRouter = require('./users-bookmarks-contributions-routes/bookmarks')
-const hideprojectsRouter = require('./users-bookmarks-contributions-routes/hideprojects')
-const contributionsRouter = require('./users-bookmarks-contributions-routes/contributions')
+const bookmarksRouter = require('./users-bookmarks-contributions-hide-routes/bookmarks');
+const hideprojectsRouter = require('./users-bookmarks-contributions-hide-routes/hideprojects');
+const contributionsRouter = require('./users-bookmarks-contributions-hide-routes/contributions');
 
 const router = express.Router();
 
-const validateLogin = [
-  check('credential')
-    .exists({ checkFalsy: true })
-    .notEmpty()
-    .withMessage('Please provide a valid username or email.'),
-  check('password')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a password.'),
-  handleValidationErrors
-];
-
-const validateSignup = [
-  check('username')
-    .exists({ checkFalsy: true })
-    .isLength({ min: 4 })
-    .withMessage('Please provide a username with at least 4 characters.'),
-  check('username')
-    .exists({ checkFalsy: true })
-    .isLength({ max: 50 })
-    .withMessage('Please provide a username with less than 51 characters.'),
-  check('email')
-    .exists({ checkFalsy: true })
-    .isLength({ min: 4 })
-    .withMessage('Please provide a email with at least 4 characters.'),
-  check('email')
-    .exists({ checkFalsy: true })
-    .isLength({ max: 50 })
-    .withMessage('Please provide a email with less than 51 characters.'),
-  check('password')
-    .exists({ checkFalsy: true })
-    .isLength({ min: 6 })
-    .withMessage('Password must be 6 characters or more.'),
-  handleValidationErrors
-];
-
-router.use('/:userId/Bookmarks', bookmarksRouter)
-router.use('/:userId/hide-project', hideprojectsRouter)
-router.use('/:userId/contributions', contributionsRouter)
+router.use('/:userId/Bookmarks', bookmarksRouter);
+router.use('/:userId/hide-project', hideprojectsRouter);
+router.use('/:userId/contributions', contributionsRouter);
 
 // Sign up
 router.post(
