+++ src/backend-src/routes/index.ts
@@ -37,7 +37,7 @@ if (process.env.NODE_ENV === 'production') {
     res.cookie('XSRF-TOKEN', req.csrfToken());
     return res.json({});
   });
-}
+};
 
 
 module.exports = router;
